Move Error exceptions to their respective modules
