+++ darkseid/archivers/__init__.py
@@ -1,8 +1,22 @@
 """Archive handling modules for different file formats."""
 
-from darkseid.archivers.archiver import Archiver
+from darkseid.archivers.archiver import (
+    Archiver,
+    ArchiverError,
+    ArchiverReadError,
+    ArchiverWriteError,
+)
 from darkseid.archivers.factory import ArchiverFactory, UnknownArchiver
 from darkseid.archivers.rar import RarArchiver
 from darkseid.archivers.zip import ZipArchiver
 
-__all__ = ["Archiver", "ArchiverFactory", "RarArchiver", "UnknownArchiver", "ZipArchiver"]
+__all__ = [
+    "Archiver",
+    "ArchiverError",
+    "ArchiverFactory",
+    "ArchiverReadError",
+    "ArchiverWriteError",
+    "RarArchiver",
+    "UnknownArchiver",
+    "ZipArchiver",
+]
